tools_dev/pdf/web/requirements.txt
flask
pypdf
pymupdf
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
